+++ doctr/datasets/datasets/tensorflow.py
@@ -7,9 +7,10 @@ import os
 from copy import deepcopy
 from typing import Any, List, Tuple
 
+import numpy as np
 import tensorflow as tf
 
-from doctr.io import read_img_as_tensor
+from doctr.io import read_img_as_tensor, tensor_from_numpy
 
 from .base import _AbstractDataset, _VisionDataset
 
@@ -21,7 +22,8 @@ class AbstractDataset(_AbstractDataset):
     def _read_sample(self, index: int) -> Tuple[tf.Tensor, Any]:
         img_name, target = self.data[index]
         # Read image
-        img = read_img_as_tensor(os.path.join(self.root, img_name), dtype=tf.float32)
+        img = tensor_from_numpy(img_name, dtype=tf.float32) if isinstance(img_name, np.ndarray) \
+            else read_img_as_tensor(os.path.join(self.root, img_name), dtype=tf.float32)
 
         return img, deepcopy(target)
